+++ packages/remix-react/index.tsx
@@ -26,11 +26,7 @@ export {
   useLoaderData,
   useActionData,
   useBeforeUnload,
-  useMatches,
-  // deprecated
-  usePendingLocation,
-  usePendingFormSubmit,
-  useRouteData
+  useMatches
 } from "./components";
 
 export type { FormMethod, FormEncType } from "./data";